Move OpenAiSpeechToTextService.java to base/src/test/java/ai/javaclaw/speech/MockSpeechToTextService and refactor OpenAiSpeechToTextService to deletate it to SpringAI
