+++ base/src/test/java/ai/javaclaw/speech/MockSpeechToTextService.java
@@ -1,12 +1,7 @@
 package ai.javaclaw.speech;
 
-import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
-import org.springframework.stereotype.Service;
-
 import java.io.InputStream;
 
-@Service
-@ConditionalOnProperty(name = "speech.provider", havingValue = "mock", matchIfMissing = true)
 public class MockSpeechToTextService implements SpeechToTextService {
 
     @Override
